Change names of Cli,Gui to Cli,Main
