+++ jflex/src/main/java/jflex/BUILD.bazel
@@ -1,9 +1,23 @@
 package(default_visibility = ["//visibility:public"])
 
-# cli
+
+# cli, gui, and ant task
+java_library(
+    name = "jflex_ui",
+    srcs = ["Main.java"],
+    deps = [
+        ":jflex_cli",
+        "//jflex/src/main/java/jflex/anttask",
+        "//jflex/src/main/java/jflex/exceptions",
+        "//jflex/src/main/java/jflex/gui",
+        "@maven//:com_google_guava_guava",
+    ],
+)
+
+# cli only
 java_library(
     name = "jflex_cli",
-    srcs = ["Main.java"],  # cli
+    srcs = ["Cli.java"],  # cli
     deps = [
         "//jflex/src/main/java/jflex/base",
         "//jflex/src/main/java/jflex/core",
@@ -18,15 +32,3 @@ java_library(
     ],
 )
 
-# cli, gui, and ant task
-java_library(
-    name = "jflex_gui",
-    srcs = ["Gui.java"],
-    deps = [
-        ":jflex_cli",
-        "//jflex/src/main/java/jflex/anttask",
-        "//jflex/src/main/java/jflex/exceptions",
-        "//jflex/src/main/java/jflex/gui",
-        "@maven//:com_google_guava_guava",
-    ],
-)
