+++ testsuite/jflex-testsuite-maven-plugin/src/main/java/jflex/maven/plugin/testsuite/ExecUtils.java
@@ -10,6 +10,7 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.List;
+import jflex.Cli;
 import jflex.core.OptionUtils;
 import jflex.exceptions.GeneratorException;
 import jflex.exceptions.SilentExit;
@@ -96,7 +97,7 @@ public class ExecUtils {
     try {
       OptionUtils.setDefaultOptions();
       Out.setOutputStream(out);
-      jflex.Main.generate(cmd);
+      Cli.generate(cmd);
       return new TestResult(out.toString(), true);
     } catch (GeneratorException e) {
       return new TestResult(out.toString(), false);
